+++ apps/runner/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @t3x-dev/runner
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @t3x-dev/core@0.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @t3x-dev/core@0.0.2
+
 ## 0.2.0
 
 ### Minor Changes