+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @t3x-dev/core
 
+## 0.0.3
+
+### Patch Changes
+
+- add skipBuiltinAuth option to createApp
+
+## 0.0.2
+
+### Patch Changes
+
+- Add frame-graph types, Delta, SemanticContent exports and fix postgres migration
+
 ## 0.2.0
 
 ### Minor Changes